+++ R/encoding.R
@@ -6,7 +6,7 @@
 #' @rdname get_encoding
 #' @name get_encoding
 #' @importFrom utils localeToCharset
-get_encoding = function(x, verbose = FALSE){
+get_encoding <- function(x, verbose = FALSE){
   enc <- unique(Encoding(x))
   if (length(enc) == 1){
     if (enc == "unknown"){
@@ -29,5 +29,3 @@ get_encoding = function(x, verbose = FALSE){
     stop("please check encoding of the input character vector - more than one encoding found")
   }
 }
-
-