+++ NEWS.md
@@ -1,3 +1,41 @@
+cwbtools 0.2.0
+==============
+
+## NEW FEATURES
+
+* The `install_corpus()` function has been reworked thoroughly. Using system directories
+  for the registry and the corpus directory is now supported. This is a prerequisite that
+  corpora can be installed outside of R packages Installing corpora within corpora is
+  not allowed by CRAN.
+* A set of new auxiliary functions (`cwb_directories()`, `cwb_registry_dir()`, 
+  `cwb_corpus_dir()`) will get the whereabouts of the registry directory and  the corpus 
+  directory.  In particular, they consider that the polmineR package may have generated a
+  temporary corpus registry, resetting the CORPUS_REGISTRY environment variable.
+* The `install_corpus()` function accepts an argument `doi` to provide a Document Object 
+  Identifier (DOI). At this stage, the DOI is assumed to be awarded by [Zenodo](https://zenodo.org/). Information available at the Zenodo site will be resolved 
+  to get the URL of a corpus tarball that can be downloaded. Upon installing a corpus
+  from Zenodo, the DOI and the version number will be written as corpus properties into 
+  the registry file.
+* To avoid removing corpora accidentally, the `corpus_install()` function will ask the user
+  for feedback if a corpus would be installed that is already present and that would be 
+  deleted or overwritten.
+* New auxiliary functions `create_cwb_directories` and `use_corpus_registry_envvar()`
+  will assist users to create the required directory structure for CWB indexed corpora.
+
+
+## MINOR IMPROVEMENTS
+
+* The default value of the argument "repo" that defines the repository for packaged corpora 
+  is now the drat repository of the PolMine GitHub account ("https://PolMine.github.io/drat/").
+
+
+## DOCUMENTATION FIXES
+
+* New R6 Roxygen documentation used for documenting the `CorpusData` class.
+* A (preliminary) vignette has been added that explains how to add a sentence annotation 
+  can be added to an existing indexed corpus.
+
+
 cwbtools 0.1.2
 ==============